+++ com.unity.render-pipelines.lightweight/Runtime/IBeforeRender.cs
@@ -1,0 +1,10 @@
+using UnityEngine.Experimental.Rendering.LightweightPipeline;
+
+namespace UnityEngine.Experimental.Rendering.LWRP
+{
+    public interface IBeforeRender
+    {
+        ScriptableRenderPass GetPassToEnqueue(RenderTextureDescriptor baseDescriptor,
+            RenderTargetHandle colorHandle, RenderTargetHandle depthHandle, ClearFlag clearFlag);
+    }
+}